requirements.txt
pyTelegramBotAPI
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pandas
flask
gspread
beautifulsoup4
lxml
ollama-python>=0.3.0
openai>=1.0.0
